requirements.txt
feedparser>=6.0.0
python-dateutil>=2.8.0
pytest>=7.0.0
trafilatura>=1.6.0
anthropic>=0.40.0
python-dotenv>=1.0.0
requests>=2.28.0
